requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
python-decouple
celery
redis
django-cors-headers
pandas
openpyxl